chore: refactor, and fix stylings
