+++ demo/app/demos/data-list/demo.component.html
@@ -1,18 +1,30 @@
-<ng-template #country let-name="name" let-desc="desc">
-  <div class="row">
-    <div>
-      <img class="responsive-image" ngSrc="https://picsum.photos/90/90" />
-    </div>
-    <div class="gutter-margin flex">
-      <div class="row">
-        <div class="row center flex">
-          <h2 class="m-0">{{ name }}</h2>
-          &nbsp;&nbsp;
-          <span class="badge">Country</span>
-        </div>
+<ng-template #kittenTemplate>
+  <div class="country-pane">
+    <div class="row">
+      <div>
+        <img class="responsive-image" src="https://picsum.photos/180/180" />
       </div>
-      <div *ngIf="desc">{{ desc }}</div>
+      <div class="gutter-margin flex">
+        <div class="row">
+          <div class="row center flex">
+            <h2 class="m-0">Kitty Cat</h2>
+            &nbsp;&nbsp;
+            <span class="badge">Pet</span>
+          </div>
+          <div class="secondary-text-color align-right">
+            <i>September 2, 2023</i><br />
+            <span>Julia</span>
+          </div>
+        </div>
+        <div>Idiosyncratic creature which makes miaow</div>
+      </div>
     </div>
+    <button
+      (click)="onDeleteCountry(id, $event)"
+      vcl-button
+      class="square transparent hide delete-button">
+      <vcl-icon icon="mdi mdi-trash-can" aria-label="Delete Country"></vcl-icon>
+    </button>
   </div>
 </ng-template>
 
@@ -22,21 +34,13 @@
 
 <vcl-data-list [(value)]="value3">
   <vcl-data-list-header>
-    <h2>Countries</h2>
+    <h2>Kittens</h2>
   </vcl-data-list-header>
   <vcl-data-list-item value="gr">
-    <ng-container
-      *ngTemplateOutlet="
-        country;
-        context: {
-          name: 'Greece',
-          desc: 'A country in southeastern Europe with thousands of islands throughout the Aegean and Ionian seas'
-        }
-      "></ng-container>
+    <ng-container *ngTemplateOutlet="kittenTemplate"></ng-container>
   </vcl-data-list-item>
   <vcl-data-list-item value="fr">
-    <ng-container
-      *ngTemplateOutlet="country; context: { name: 'France' }"></ng-container>
+    <ng-container *ngTemplateOutlet="kittenTemplate"></ng-container>
   </vcl-data-list-item>
   <vcl-busy-indicator class="p-3 w-100p"></vcl-busy-indicator>
 </vcl-data-list>
