+++ demo/app/demos/data-list/list-add-remove/create-kitten.component.ts
@@ -5,7 +5,7 @@ import { ComponentLayerRef } from '@vcl/ng-vcl';
   selector: 'app-create-country',
   template: `
     <vcl-panel-dialog [showCloseButton]="true" (close)="close()">
-      <vcl-panel-title>Add Country</vcl-panel-title>
+      <vcl-panel-title>Add Kitten</vcl-panel-title>
 
       <vcl-form-control-group>
         <vcl-label>Name</vcl-label>
@@ -21,13 +21,6 @@ import { ComponentLayerRef } from '@vcl/ng-vcl';
         </vcl-input-field>
       </vcl-form-control-group>
 
-      <vcl-form-control-group>
-        <vcl-label>Code</vcl-label>
-        <vcl-input-field>
-          <input vclInput [(ngModel)]="code" />
-        </vcl-input-field>
-      </vcl-form-control-group>
-
       <button (click)="save()" vcl-button vclPanelFooterButton>Save</button>
 
       <button (click)="close()" vcl-button vclPanelFooterButton class="transparent outline">
@@ -39,7 +32,6 @@ import { ComponentLayerRef } from '@vcl/ng-vcl';
 export class CreateCountryComponent {
   name = '';
   description = '';
-  code = '';
 
   constructor(private layer: ComponentLayerRef) {}
 
@@ -48,7 +40,6 @@ export class CreateCountryComponent {
       value: {
         name: this.name,
         description: this.description,
-        code: this.code,
       },
     });
   }
