+++ demo/app/demos/data-list/list-add-remove/data-list-add-remove-demo.component.scss
@@ -9,7 +9,7 @@
 .delete-button {
   position: absolute;
   right: 0;
-  top: 0;
+  top: 2em;
 
   .icon {
     font-size: 1.5em;
